src/app/dog-list/dog-list.component.spec.ts
import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';

import { DogListComponent } from './dog-list.component';
import { DogDetailComponent } from '../dog-detail/dog-detail.component';

describe('DogListComponent', () => {
  let component: DogListComponent;
  let fixture: ComponentFixture<DogListComponent>;
  let html;

  beforeEach(waitForAsync(() => {
    TestBed.configureTestingModule({
      declarations: [ DogListComponent, DogDetailComponent ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(DogListComponent);
    component = fixture.componentInstance;
    html = fixture.nativeElement;
    fixture.autoDetectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should render a list of dogs', () => {
    expect(component.dogs).toBeDefined()
    expect(html.querySelector('dog'))
  });

  it('should render list of favorite dogs', () => {
    spyOn(component, 'loadFavorites');
    const favButton = html.querySelector('#favorites');
    favButton.click();
    expect(component.loadFavorites).toHaveBeenCalled();
  });
});
